⚡️[#143] 싱글톤 패턴 적용
